Add "expiry" configuration setting to exclude old messages
Added an "expiry" setting in the config file.  Messages that are older than this setting will be excluded from the sensor.

Note: I was unable to make this an optional setting.  I suspect that the cv.time_period_dict type cannot be optional.  

This makes the proposed upgrade a breaking change!